gtag type 설정
