+++ meezzle-front/package.json
@@ -26,6 +26,7 @@
         "uuid": "^9.0.0"
     },
     "devDependencies": {
+        "@types/gtag.js": "^0.0.12",
         "@types/node": "18.8.3",
         "@types/react": "18.0.21",
         "@types/react-datepicker": "^4.8.0",
