+++ meezzle-front/pages/_app.tsx
@@ -6,6 +6,7 @@ import Head from "next/head";
 import * as gtag from "../utils/gtag";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
+import { GA_TRACKING_ID } from "../utils/gtag";
 
 import {
     DehydratedState,
@@ -88,6 +89,23 @@ function MyApp({
                             type="image/x-icon"
                             href="/logo.ico"
                         />
+                         {/* Global Site Tag (gtag.js) - Google Analytics */}
+          <script
+            async
+            src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
+          />
+          <script
+            dangerouslySetInnerHTML={{
+              __html: `
+              window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              gtag('config', '${GA_TRACKING_ID}', {
+                page_path: window.location.pathname,
+              });
+          `
+            }}
+          />
                     </Head>
                     <GlobalStyle />
                     <Component {...pageProps} />
