+++ meezzle-front/utils/gtag.ts
@@ -1,5 +1,9 @@
 export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_GA_TRACKING_ID;
-
+declare global {
+  interface Window {
+    gtag: any
+  }
+}
 // https://developers.google.com/analytics/devguides/collection/gtagjs/pages
 export const pageview = (url: URL) => {
   window.gtag("config", GA_TRACKING_ID, {
